+++ src/map/map_layer.cpp
@@ -130,8 +130,8 @@ void CMapLayer::RegenerateForestTile(const Vec2i &pos)
 	//  FIXME: a better looking result would be fine
 	//    Allow general updates to any tiletype that regrows
 
-	const unsigned long permanent_occupied_flag = (MapFieldWall | MapFieldUnpassable | MapFieldBuilding);
-	const unsigned long occupied_flag = (permanent_occupied_flag | MapFieldLandUnit | MapFieldItem);
+	const uint16_t permanent_occupied_flag = (MapFieldWall | MapFieldUnpassable | MapFieldBuilding);
+	const uint16_t occupied_flag = (permanent_occupied_flag | MapFieldLandUnit | MapFieldItem);
 	
 	if ((mf.GetFlags() & permanent_occupied_flag)) { //if the tree tile is permanently occupied by buildings and the like, reset the regeneration process
 		mf.Value = 0;
